feat(autoware_ptv3): split the backbone and head (autowarefoundation/autoware_universe#12655)
* feat(autoware_ptv3): add multi-head support (segmentation3d and detection3d)


* fix(autoware_ptv3): cspell


* fix(autoware_ptv3): typo


* fix(autoware_ptv3): centerhead dims order


* feat(autoware_ptv3): remove center_head support for now


* refactor(autoware_ptv3): PR split - drop det3d head


* feat(autoware_ptv3): cleanup


* feat(autoware_ptv3): explicit types for all tensors


* feat(autoware_ptv3): pre-commit


* feat(autoware_ptv3): restore old sources


* fix(autoware_ptv3): use of member initializer lists


* fix(autoware_ptv3): remove redundant initialization


* fix(autoware_ptv3): wrong type


* fix(autoware_ptv3): redundant sync


* fix(autoware_ptv3): guard pointer


* feat(autoware_ptv3): update model params


---------

+++ perception/autoware_ptv3/include/autoware/ptv3/preprocess/preprocess_kernel.hpp
@@ -34,7 +34,7 @@ public:
 
   std::size_t generateFeatures(
     const void * input_data, CloudFormat input_format, unsigned int num_points,
-    float * voxel_features, std::int64_t * voxel_coords, std::int64_t * voxel_hashes,
+    float * voxel_features, std::int32_t * voxel_coords, std::int64_t * voxel_hashes,
     void * compact_points, float * reconstruction_features, void * cropped_source_points,
     std::int64_t * inverse_map, std::size_t * num_cropped_points);
